attempt_lint_1/requirements.txt
torch
PyYAML
sentencepiece
nltk
numpy